Extract a Document component to hold the markdown-presenting code
