+++ src/contexts/current-markdown.ts
@@ -1,4 +1,5 @@
 import { Dispatch, createContext } from "react";
-import type { MarkdownAction } from '../reducers/markdown-actions';
+import type { MarkdownAction, MarkdownPayload } from '../reducers/markdown-actions';
 
 export const CurrentMarkdownDispatchContext = createContext<Dispatch<MarkdownAction> | null>(null);
+export const CurrentMarkdownStateContext = createContext<MarkdownPayload | null>(null);
